Extract app config and DI container
